autenticação atualizada

+++ app/Http/Controllers/CategoriasController.php
@@ -22,20 +22,48 @@ class CategoriasController extends Controller
     public function createView( Request $request)
     {
 
-        return view('categorias.create');
+        session_start();
+        if(isset($_SESSION['autenticado']) && $_SESSION['autenticado'] == true) {
+            if($_SESSION['administrador'] == true){
+                return view('categorias.create');
+
+            }else {
+                flash('Você não tem permissão para Criar novas categorias')->error();
+                return redirect()->back();
+            }
+        }else {
+            flash('Você não tem permissão para Criar novas categorias')->error();
+            return redirect()->back();
+        }
     }
 
     public function updateView( Request $request , $categoriaId)
     {
 
-        $categoria = Categoria::find($categoriaId);
 
-        if(!$categoria)
-        {
-            flash('categoria nao encontrada')->error();
+        session_start();
+        if(isset($_SESSION['autenticado']) && $_SESSION['autenticado'] == true) {
+            if($_SESSION['administrador'] == true){
+                        
+                $categoria = Categoria::find($categoriaId);
+
+                if(!$categoria)
+                {
+                    flash('categoria nao encontrada')->error();
+                    return redirect()->back();
+                }
+                return view('categorias.update')->with(compact('categoria'));
+
+            }else {
+                flash('Você não tem permissão para Atualizar as categorias')->error();
+                return redirect()->back();
+            }
+    
+        }else {
+            flash('Você não tem permissão para Atualizar as categorias')->error();
             return redirect()->back();
         }
-        return view('categorias.update')->with(compact('categoria'));
+
     }
 
 
@@ -66,28 +94,39 @@ class CategoriasController extends Controller
     public function delete( Request $request, $categoriaId)
     {
 
-        try{
-            $categoria = Categoria::find($categoriaId);
 
-            if(!$categoria){
-                throw new Exception('Server Error : categoria nao encontrada');
-            }
-            
+        session_start();
+        if(isset($_SESSION['autenticado']) && $_SESSION['autenticado'] == true) {
+            if($_SESSION['administrador'] == true){
+                try{
+                    $categoria = Categoria::find($categoriaId);
 
-            $deletedCategory = $categoria->delete();
+                    if(!$categoria){
+                        throw new Exception('Server Error : categoria nao encontrada');
+                    }
+                    
 
-            if($deletedCategory)
-            {
-            
-                flash('Categoria deletada com sucesso')->success();
+                    $deletedCategory = $categoria->delete();
+
+                    if($deletedCategory)
+                    {
+                    
+                        flash('Categoria deletada com sucesso')->success();
+                        return redirect()->back();
+                    }
+                }catch(Exception $e)
+                {
+                    flash($e->getMessage())->error();
+                    return redirect()->back();
+                }
+            }else {
+                flash('Você não tem permissão para Deletar categorias')->error();
                 return redirect()->back();
             }
-        }catch(Exception $e)
-        {
-            flash($e->getMessage())->error();
+        }else {
+            flash('Você não tem permissão para Deletar categorias')->error();
             return redirect()->back();
         }
-
     }
 
     public function update( Request $request, $categoriaId)
