+++ app/Http/Controllers/ProdutosController.php
@@ -12,13 +12,23 @@ class ProdutosController extends Controller
 
     public function createView( Request $request)
     {
-        $categorias = Categoria::all();
-        return view('produtos.create')->with(compact('categorias'));
+        session_start();
+        if(isset($_SESSION['autenticado']) && $_SESSION['autenticado'] == true) {
+            if($_SESSION['administrador'] == true){
+                $categorias = Categoria::all();
+                return view('produtos.create')->with(compact('categorias'));
+            }else {
+                flash('Você não tem permissão para Criar novos produtos')->error();
+                return redirect()->back();
+            }
+        }else {
+            flash('Você não tem permissão para Criar novos produtos')->error();
+            return redirect()->back();
+        }
+        
     }
     public function show( Request $request)
     {
-
-      
         $produtos = Produto::all();
         return view('produtos.show')->with(compact('produtos'));
     }
@@ -26,23 +36,37 @@ class ProdutosController extends Controller
 
     public function updateView( Request $request , $produtoId)
     {
+        session_start();
+        if(isset($_SESSION['autenticado']) && $_SESSION['autenticado'] == true) {
+            if($_SESSION['administrador'] == true){
 
-        $produto = Produto::find($produtoId);
-        
-        if(!$produto)
-        {
-            flash('produto não encontrado')->error();
+                        
+                $produto = Produto::find($produtoId);
+                
+                if(!$produto)
+                {
+                    flash('produto não encontrado')->error();
+                    return redirect()->back();
+                }
+
+                $categorias = Categoria::all();
+
+                if(!$categorias)
+                {
+                    flash('Categorias não encontradas')->error();
+                    return redirect()->back();
+                }
+                return view('produtos.update')->with(compact('produto', 'categorias'));
+
+            }else {
+                flash('Você não tem permissão para atualizar os produtos')->error();
+                return redirect()->back();
+            }
+        }else {
+            flash('Você não tem permissão para atualizar os produtos')->error();
             return redirect()->back();
         }
 
-        $categorias = Categoria::all();
-
-        if(!$categorias)
-        {
-            flash('Categorias não encontradas')->error();
-            return redirect()->back();
-        }
-        return view('produtos.update')->with(compact('produto', 'categorias'));
     }
 
 
@@ -76,29 +100,42 @@ class ProdutosController extends Controller
     public function delete( Request $request, $produtoId)
     {
 
-        try{
-            $produto = Produto::find($produtoId);
-
-            if(!$produto){
-                throw new Exception('Server Error : produto não encontrado');
-            }
-            
-
-            $deletedProduto = $produto->delete();
-
-            if($deletedProduto)
-            {
-            
-                flash('Produto deletada com sucesso')->success();
-                return redirect()->back();
+        session_start();
+        if(isset($_SESSION['autenticado']) && $_SESSION['autenticado'] == true) {
+            if($_SESSION['administrador'] == true){
+                try{
+                    $produto = Produto::find($produtoId);
+        
+                    if(!$produto){
+                        throw new Exception('Server Error : produto não encontrado');
+                    }
+                    
+        
+                    $deletedProduto = $produto->delete();
+        
+                    if($deletedProduto)
+                    {
+                    
+                        flash('Produto deletada com sucesso')->success();
+                        return redirect()->back();
+                    }else {
+                        throw new Exception('Server Error : produto não deletado, erro desconhecido');
+                    }
+                }catch(Exception $e)
+                {
+                    flash($e->getMessage())->error();
+                    return redirect()->back();
+                }
             }else {
-                throw new Exception('Server Error : produto não deletado, erro desconhecido');
+                flash('Você não tem permissão para Deletar produtos')->error();
+                return redirect()->back();
             }
-        }catch(Exception $e)
-        {
-            flash($e->getMessage())->error();
+        }else {
+            flash('Você não tem permissão para Deletar produtos')->error();
             return redirect()->back();
         }
+
+        
 
     }
